Integrate NUITranslator in UIFormat
